AB#5157 Fix: createTimeRangeArchive accepts a window period longer than 24 days
periodMs was declared Int, so every value above Int32.MaxValue was refused by variable
coercion with "Unable to convert '2592000000' to 'Int'". That is not an edge case: the Studio's
own coarsest preset, "1 month (approx. 30d)", sends exactly 2 592 000 000 ms, and a 92-day
legacy archive could therefore only be created through ImportRt (AB#5157 review).

The field is now Long and the input DTO carries long?, which is what every other millisecond
field on this surface already uses — createRollupArchive's bucketSizeMs, watermarkLagMs and
carryLookbackMs are all Long, and the SDK's ArchiveSchemaDto.PeriodMs is long?. The value still
flows into TimeSpan.FromMilliseconds unchanged.

Tests: the periodMs field is pinned to LongGraphType, and the Long scalar accepts the month
preset the Studio sends.

+++ src/AssetRepositoryServices/GraphQL/Types/Inputs/CreateTimeRangeArchiveInputDto.cs
@@ -13,7 +13,7 @@ internal sealed class CreateTimeRangeArchiveInputDto
     public string? RtWellKnownName { get; set; }
     public string TargetCkTypeId { get; set; } = string.Empty;
     public List<ArchiveColumnSpecInputDto> Columns { get; set; } = new();
-    public int? PeriodMs { get; set; }
+    public long? PeriodMs { get; set; }
 }
 
 /// <summary>
